+++ index.js
@@ -11,12 +11,12 @@
  */
 
 var cookie = require('cookie');
+var crc = require('crc').crc32;
 var debug = require('debug')('express-session');
 var deprecate = require('depd')('express-session');
 var parseUrl = require('parseurl');
 var uid = require('uid-safe').sync
   , onHeaders = require('on-headers')
-  , crc32 = require('buffer-crc32')
   , signature = require('cookie-signature')
 
 var Session = require('./session/session')
@@ -155,8 +155,9 @@ function session(options){
     // ensure secret is available or bail
     if (!secret) next(new Error('`secret` option required for sessions'));
 
-    var originalHash
-      , originalId;
+    var originalHash;
+    var originalId;
+    var savedHash;
 
     // expose store
     req.sessionStore = store;
@@ -270,16 +271,15 @@ function session(options){
         return _end.call(res, chunk, encoding);
       }
 
-      req.session.resetMaxAge();
+      // touch session
+      req.session.touch();
 
       if (shouldSave(req)) {
-        debug('saving');
         req.session.save(function onsave(err) {
           if (err) {
             defer(next, err);
           }
 
-          debug('saved');
           writeend();
         });
 
@@ -294,11 +294,35 @@ function session(options){
       store.generate(req);
       originalId = req.sessionID;
       originalHash = hash(req.session);
+      wrapmethods(req.session);
+    }
+
+    // wrap session methods
+    function wrapmethods(sess) {
+      var _save = sess.save;
+
+      function save() {
+        debug('saving %s', this.id);
+        savedHash = hash(this);
+        _save.apply(this, arguments);
+      }
+
+      Object.defineProperty(sess, 'save', {
+        configurable: true,
+        enumerable: false,
+        value: save,
+        writable: true
+      });
     }
 
     // check if session has been modified
     function isModified(sess) {
-      return originalHash != hash(sess) || originalId != sess.id;
+      return originalId !== sess.id || originalHash !== hash(sess);
+    }
+
+    // check if session has been saved
+    function isSaved(sess) {
+      return originalId === sess.id && savedHash === hash(sess);
     }
 
     // determine if session should be destroyed
@@ -308,9 +332,9 @@ function session(options){
 
     // determine if session should be saved to store
     function shouldSave(req) {
-      return sessionId != req.sessionID
-        ? saveUninitializedSession || isModified(req.session)
-        : resaveSession || isModified(req.session);
+      return !saveUninitializedSession && cookieId !== req.sessionID
+        ? isModified(req.session)
+        : !isSaved(req.session)
     }
 
     // determine if cookie should be set on response
@@ -339,25 +363,32 @@ function session(options){
       // error handling
       if (err) {
         debug('error %j', err);
-        if ('ENOENT' == err.code) {
-          generate();
-          next();
-        } else {
+
+        if (err.code !== 'ENOENT') {
           next(err);
+          return;
         }
+
+        generate();
       // no session
       } else if (!sess) {
         debug('no session found');
         generate();
-        next();
       // populate req.session
       } else {
         debug('session found');
         store.createSession(req, sess);
         originalId = req.sessionID;
         originalHash = hash(sess);
-        next();
+
+        if (!resaveSession) {
+          savedHash = originalHash
+        }
+
+        wrapmethods(req.session);
       }
+
+      next();
     });
   };
 };
@@ -448,8 +479,10 @@ function getcookie(req, name, secret) {
  */
 
 function hash(sess) {
-  return crc32.signed(JSON.stringify(sess, function(key, val){
-    if ('cookie' != key) return val;
+  return crc(JSON.stringify(sess, function (key, val) {
+    if (key !== 'cookie') {
+      return val;
+    }
   }));
 }